+++ example/android/app/src/main/java/com/fuadarradhi/datetime_setting_example/MainActivity.java
@@ -1,8 +1,0 @@
-package com.fuadarradhi.datetime_setting_example;
-
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.plugins.firebase.core.FirebaseCorePlugin;
-
-public class MainActivity extends FlutterActivity {
-}