+++ lib/datetime_setting.dart
@@ -12,4 +12,8 @@ class DatetimeSetting {
   static Future<bool> timeZoneIsAuto() async {
     return await _channel.invokeMethod('timeZoneIsAuto');
   }
+
+  static Future<void> openSetting() async {
+    return await _channel.invokeMethod('openSetting');
+  }
 }