+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 ## 2.0.0
 
 * null safety.
+
+## 2.0.1
+
+* v2 embedding (thanks @jrmessias)
+
+## 2.0.2
+
+* Get back "openSetting" function